создано отдельное приложение отзывы
обновлены все связи, сериализатор и вью
