+++ eatpoint/api/serializers/establishments.py
@@ -5,7 +5,6 @@ from drf_spectacular.utils import (
     extend_schema_field,
 )
 
-
 from rest_framework import serializers
 from phonenumber_field.serializerfields import PhoneNumberField
 
@@ -14,12 +13,10 @@ from core.services import days_available
 from core.validators import validate_uniq
 from establishments.models import (
     Establishment,
-    OwnerResponse,
     WorkEstablishment,
     Kitchen,
     ZoneEstablishment,
     Favorite,
-    Review,
     Service,
     SocialEstablishment,
     ImageEstablishment,
@@ -28,7 +25,7 @@ from establishments.models import (
     Event,
 )
 from reservation.models import Availability
-from users.models import User
+from reviews.models import Review
 
 
 class KitchenSerializer(serializers.ModelSerializer):
@@ -436,74 +433,6 @@ class EstablishmentEditSerializer(serializers.ModelSerializer):
             instance,
             context={"request": self.context.get("request")},
         ).data
-
-
-class SmallUserSerializer(serializers.ModelSerializer):
-    """Сериализация данных: Данные пользователя для отзывов"""
-
-    class Meta:
-        model = User
-        fields = [
-            "first_name",
-            "last_name",
-            "role",
-        ]
-
-
-class ReviewSerializer(serializers.ModelSerializer):
-    """Сериализация данных: Отзывы"""
-
-    establishment = serializers.SlugRelatedField(
-        slug_field="name",
-        read_only=True,
-    )
-    author = SmallUserSerializer(read_only=True)
-    response = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = "__all__"
-        model = Review
-
-    def get_response(self, obj):
-        owner_response = obj.owner_responses.first()
-        if owner_response:
-            return OwnerResponseSerializer(owner_response).data
-        else:
-            return None
-
-    def validate(self, data):
-        """Проверка на уникальность отзыва"""
-        if self.context["request"].method == "POST":
-            establishment = self.context["view"].kwargs.get("establishment_id")
-            user = self.context["request"].user
-
-            is_booking_confirmed = user.reservationhistory.filter(
-                establishment=establishment, status=True
-            ).exists()
-            if not is_booking_confirmed:
-                raise serializers.ValidationError(
-                    "Отзыв недоступен из-за отсутствия подтвержденной брони"
-                )
-
-            if Review.objects.filter(
-                author=user, establishment=establishment
-            ).exists():
-                raise serializers.ValidationError(
-                    "Нельзя оставить повторный отзыв на одно заведение"
-                )
-        return data
-
-
-class OwnerResponseSerializer(serializers.ModelSerializer):
-    """
-    Сериализация данных: Ответ владельца заведения на отзыв.
-    Для получения информации о самом отзыве в ответе API при
-    создании ответа владельца на отзыв, добавлен 'review'.
-    """
-
-    class Meta:
-        model = OwnerResponse
-        fields = ["id", "text", "created"]
 
 
 class EventSerializer(serializers.ModelSerializer):
