+++ eatpoint/api/urls.py
@@ -4,7 +4,6 @@ from rest_framework.routers import DefaultRouter
 from api.views.analytics import AnalyticsViewSet, AnalyticsListViewSet
 from api.views.code_generate import SendSMSCode, VerifySMSCode
 from api.views.establishments import (
-    OwnerResponseCreateView,
     ZoneViewSet,
     CityViewSet,
     EstablishmentBusinessViewSet,
@@ -20,6 +19,7 @@ from api.views.reservation import (
     ReservationsRestorateurListViewSet,
     AvailabilityViewSet,
 )
+from api.views.reviews import OwnerResponseCreateView, ReviewViewSet
 from api.views.users import (
     SignUp,
     ConfirmCodeView,
@@ -31,7 +31,6 @@ from api.views.users import (
 )
 from api.views.establishments import (
     EstablishmentViewSet,
-    ReviewViewSet,
     KitchenViewSet,
     ServicesViewSet,
     TypeEstViewSet,
